+++ developer/utils/option_parser.py
@@ -1,0 +1,20 @@
+# Copyright (c) 2017 Shotgun Software Inc.
+#
+# CONFIDENTIAL AND PROPRIETARY
+#
+# This work is provided "AS IS" and subject to the Shotgun Pipeline Toolkit
+# Source Code License included in this distribution package. See LICENSE.
+# By accessing, using, copying or modifying this work you indicate your
+# agreement to the Shotgun Pipeline Toolkit Source Code License. All rights
+# not expressly granted therein are reserved by Shotgun Software Inc.
+
+import optparse
+
+
+class OptionParserLineBreakingEpilog(optparse.OptionParser):
+    """
+    Subclassed version of the option parser that doesn't
+    swallow white space in the epilog
+    """
+    def format_epilog(self, formatter):
+        return self.epilog
